Collision detection moved to GameObject. Optimizations.
