+++ src/js/Shot.js
@@ -19,30 +19,7 @@ class Shot extends GameObject {
       return;
     }
 
-    let shotStart = this.position.clone();
-    let shotEnd = shotStart.clone();
-    let direction = new THREE.Vector3(0, 0, 1);
-    direction.applyQuaternion(this.quaternion);
-    direction.multiplyScalar(this.velocity * delta);
-    shotEnd.add(direction);
-    let denominator = shotStart.distanceTo(shotEnd);
-
-    let hitObject;
-    let hitDistance = Infinity;
-    for (let shootable of COLLIDABLES) {
-      if (shootable === this.owner) continue;
-      let shootableCenter = shootable.position;
-      if (this.position.distanceTo(shootableCenter) > denominator + shootable.hitRadius) continue;
-      let a1 = new THREE.Vector3().subVectors(shootableCenter, shotStart);
-      let a2 = new THREE.Vector3().subVectors(shootableCenter, shotEnd);
-      let radius = a1.cross(a2).length() / denominator;
-      if (radius > shootable.hitRadius) continue;
-      let hitDistance2 = this.position.distanceTo(shootableCenter);
-      if (hitDistance2 < hitDistance) {
-        hitDistance = hitDistance2;
-        hitObject = shootable;
-      }
-    }
+    let hitObject = this.checkCollision(this.quaternion, this.velocity * delta);
     if (hitObject) {
       this.remove();
       hitObject.dealDamage(this.damage);
